patch for apple m1

+++ util/crc32c_arm64.h
@@ -7,12 +7,13 @@
 #define UTIL_CRC32C_ARM64_H
 
 #include <inttypes.h>
+#include <cstddef>
 
 #if defined(__aarch64__) || defined(__AARCH64__)
 #ifdef __ARM_FEATURE_CRC32
 #define HAVE_ARM64_CRC
 #include <arm_acle.h>
-extern uint32_t crc32c_arm64(uint32_t crc, unsigned char const *data, unsigned len);
+extern uint32_t crc32c_arm64(uint32_t crc, unsigned char const *data, size_t len);
 extern uint32_t crc32c_runtime_check(void);
 #endif
 #endif
